refactor(#72): make table info component reusable
